Exposes many GPIOs for the DCM
This is part of the DCM bringup. I noticed that the TPM GPIOs were
not showing up in the Kernel output, and there were an assortment
of other GPIOs also not showing up, so I took some time to fix
as many as I could.
The notable exception was overtemp_shdn, which I couldn't get working,
and will have to wait until later.