Enables SS power monitor chips
These were disabled early in development of the SmartSense, along with
two other chips (comment added) which were only on the dev-kit (judging
from the schematic). It is assumed that they were disabled by accident.

Years later, access to these chips was requested by the CV team, so that
they could do the same power modelling that they did on the CTM. This
commit enables the sensors for them.

Fixes HW-3710